requirements.txt
discord.py==1.7.1
PyNaCl==1.4.0
async-timeout==3.0.1
asyncio==3.4.3
gTTS==2.2.2
pixiv-api==1.0.0
xmltodict==0.12.0
pathlib==1.0.1
arrow==1.0.3
beautifulsoup4==4.9.3
Pillow==8.2.0
tabula-py==2.3.0
lavalink